pyproject.toml
[project]
name = "pendulum-simulation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gymnasium>=1.0.0",
    "matplotlib>=3.10.0",
    "neat-python>=0.92",
    "numpy>=2.0.2",
    "pygame>=2.6.1",
    "pyqt6>=6.8.0",
    "tensorflow[and-cuda]>=2.18.0",
    "tf-keras>=2.18.0",
]
